pyproject.toml
[project]
name = "tc-01"
version = "0.1.0"
description = ""
authors = [
    {name = "Valterlan Dantas",email = "f.valterlan@gmail.com"}, 
    {name = "Matheus Silvestre",email = "matheus@silvestre.dev"} 
]
readme = "README.md"
requires-python = ">=3.11" #alterei para 3.11 para ter mais compatibilidade num futuro deploy.
dependencies = [           #alterado para melhor compatibilidade
    "requests>=2.32.0,<3.0.0",
    "beautifulsoup4>=4.12,<5.0",
    "fastapi>=0.112,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "PyJWT>=2.9,<3.0"
]

[tool.poetry]
packages = [{include = "tc_01", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
